+++ packages/stream_chat/test/src/core/models/message_state_test.dart
@@ -270,7 +270,10 @@ void main() {
     test(
       'MessageState.sendingFailed should create a MessageFailed instance with SendingFailed state',
       () {
-        const messageState = MessageState.sendingFailed;
+        final messageState = MessageState.sendingFailed(
+          skipPush: false,
+          skipEnrichUrl: false,
+        );
         expect(messageState, isA<MessageFailed>());
         expect((messageState as MessageFailed).state, isA<SendingFailed>());
       },
@@ -279,7 +282,7 @@ void main() {
     test(
       'MessageState.updatingFailed should create a MessageFailed instance with UpdatingFailed state',
       () {
-        const messageState = MessageState.updatingFailed;
+        final messageState = MessageState.updatingFailed(skipEnrichUrl: false);
         expect(messageState, isA<MessageFailed>());
         expect((messageState as MessageFailed).state, isA<UpdatingFailed>());
       },